closes #466 - added if file exists around unlink to prevent 500 error

+++ lib/TAN/Controller/Minify.pm
@@ -60,7 +60,9 @@ sub index: Path{
                     flock( $fh, LOCK_UN );
                 }
                 close( $fh );
-                unlink( '/tmp/tan_old_js_clean' ) || die "failed to unlink '/tmp/tan_old_js_clean' $!";
+                if ( -e '/tmp/tan_old_js_clean' ){
+                    unlink( '/tmp/tan_old_js_clean' ) || die "failed to unlink '/tmp/tan_old_js_clean' $!";
+                }
             }
         }
